requirements.txt
cryptography==3.4.7
Flask==2.0.1
PyJWT==2.1.0
